refactor: centralize migration markers and shared constants
Move the KATAMARAN_* log-marker spellings into internal/migration
markers.go so the source emitters and the orchestrator scrapers cannot
drift apart, add one migrationCapabilities constructor used by both
ends of the QEMU handshake, and route every downtime-limit validation
through a single MaxDowntimeMS constant.

Also group transformCmdline's six positional substitutions into a
cmdlineRewrite struct that honors the configured tap interface, widen
the pid fields in MigrationMeta and the factory gRPC surface to int64,
and factor out the progress-log threshold checks.

+++ internal/factory/server.go
@@ -165,7 +165,7 @@ func (s *Server) GetBaseVM(ctx context.Context, _ *emptypb.Empty) (*cachepb.Grpc
 	return &cachepb.GrpcVM{
 		Id:         state.ID,
 		Hypervisor: state.HypervisorState,
-		ProxyPid:   int64(state.VirtiofsdPid),
+		ProxyPid:   state.VirtiofsdPid,
 		Cpu:        state.CPU,
 		Memory:     state.Memory,
 	}, nil
@@ -178,7 +178,7 @@ func (s *Server) Status(_ context.Context, _ *emptypb.Empty) (*cachepb.GrpcStatu
 	vms := make([]*cachepb.GrpcVMStatus, n)
 	for i, st := range s.queue {
 		vms[i] = &cachepb.GrpcVMStatus{
-			Pid:    int64(st.QEMUPid),
+			Pid:    st.QEMUPid,
 			Cpu:    st.CPU,
 			Memory: st.Memory,
 		}
